+++ src/applications/vre/27-8832/config/chapters/claimant-information/staticClaimantComponent.jsx
