+++ src/applications/vre/27-8832/config/chapters/claimant-information/staticClaimantInformation.js
@@ -1,5 +1,5 @@
 import ClaimantInformationComponent from './staticClaimantComponent.jsx';
-import { claimantStaticInformation } from '../../utilities';
+import { claimantStaticInformation } from '../../utilities.js';
 
 export const schema = claimantStaticInformation;
