+++ src/applications/vre/27-8832/manifest.json
