+++ src/applications/vre/27-8832/reducers/index.js
